fix: review pass on LlmCompaction — defaults, provenance, cost visibility
Review of the received implementation against main. The design held up; the
defaults and two claims did not.

Blocking fixes:

- retain_tail_tokens defaulted to a flat 20_000, which made the whole strategy
  a SILENT no-op on any budget under ~33k: summarization is first attempted at
  trigger_ratio * budget, and if the tail alone consumes all the history that
  exists at that point, choose_cut finds no split and returns None with nothing
  logged. Simulating the loop's per-turn compact() call at a 20k budget over
  120 turns produced output byte-identical to DefaultCompaction — same 6 cache
  breaks on the same turns, zero summaries, zero requests. The default now
  derives from the budget, min(20_000, budget / 4), recomputed per call so it
  tracks the loop's calibrated budget; an impossible configuration warns once
  instead of failing quietly. Both covered by tests.

- temperature is left at None. There is no temperature quirk flag in the compat
  matrix, openai_compat forwards the value verbatim, and the newest reasoning
  models reject sampling parameters outright.

- Constructor is now from_config(ModelConfig), mirroring Agent::from_config, so
  provider, model id and key cannot drift apart and base_url / headers / compat
  flags reach the request. from_provider stays for test doubles and says what it
  gives up. model_config is threaded through StreamConfig.

- Fingerprint hashes serialized bytes rather than (timestamp, token estimate).
  The old fold could not see a same-length, same-timestamp rewrite; a test now
  performs exactly that edit. Its doc rationale was also wrong: transform_context
  cannot invalidate the head (the loop runs it on a clone), whereas
  Agent::replace_messages and this strategy's own fallback can.

- The summary covered [0, cut) while the splice also re-emitted [0, head_end)
  verbatim, putting the head in context twice. It now covers [head_end, cut).

- StreamConfig built via new() + field mutation, per its documented convention.

Also in this pass:

- AgentEvent::ContextCompacted + CompactionMethod, emitted on both the splice
  and the deterministic fallback, carrying messages/tokens before and after and
  the summarization request's own Usage and dollar cost — the numbers needed to
  price this strategy against DefaultCompaction rather than guess. Snapshot case
  added; EVENT_VARIANT_COUNT 12 -> 13.
- Module renamed compaction_llm -> llm_compaction (house noun-order).
- Module docs rewritten. The cache-savings framing was wrong: breaks are a wash
  against DefaultCompaction (6 vs 6 over 120 turns at 20k, 6 vs 5 over 600 at
  100k), both rewrite only on a budget crossing. Docs now claim retention
  quality at no added loop latency, state the per-summarization token cost, and
  record that compact_target_ratio / compact_headroom_turns do not apply.

Deferred, deliberately: consuming the loop's adapted compaction target, cache
breakpoint parity across providers, SharedState truncation-retrieval.

193 existing + 9 new = 202 lib tests.
